backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
pillow>=10.4.0
torch>=2.5.1
torchvision>=0.20.1
numpy>=1.26.4
opencv-python>=4.10.0.84
pydantic>=2.9.0
aiofiles>=24.1.0
timm>=0.9.0
scikit-learn>=1.3.0

# Optional but recommended for full GAT functionality
# torch-geometric>=2.4.0
